requirements.txt
groq
uvicorn
fastapi
python-dotenv